NotificationsFragment that actually works. But UGLY.
