+++ app/src/main/java/edu/outside2154/gamesense/activity/NavActivity.kt
@@ -12,6 +12,7 @@ import android.view.MenuItem
 import edu.outside2154.gamesense.R
 import edu.outside2154.gamesense.fragment.ChecklistFragment
 import edu.outside2154.gamesense.fragment.HomeFragment
+import edu.outside2154.gamesense.fragment.NotificationsFragment
 import edu.outside2154.gamesense.fragment.SettingsFragment
 import edu.outside2154.gamesense.util.transact
 
@@ -51,6 +52,7 @@ class NavActivity : AppCompatActivity() {
             R.id.nav_home_fragment -> HomeFragment()
             R.id.nav_settings_fragment -> SettingsFragment()
             R.id.nav_checklist_fragment -> ChecklistFragment()
+            R.id.nav_notification_fragment -> NotificationsFragment()
             else -> Fragment()  // TODO: replace with a 404 fragment
         }
